// apps/web/src/api/models/documentViewerContextResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * MESA Law API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentResponse } from './documentResponse';
import type { DocumentRevisionViewerResponse } from './documentRevisionViewerResponse';
import type { ParsedDocumentViewerResponse } from './parsedDocumentViewerResponse';

export interface DocumentViewerContextResponse {
  document: DocumentResponse;
  parsed_document?: ParsedDocumentViewerResponse | null;
  revision?: DocumentRevisionViewerResponse | null;
}